requirements.txt
pandas
numpy
matplotlib
seaborn
tensorflow
statsmodels
scikit-learn
tqdm
contextlib
logging
times
tensorflow
pyserial